OrderPage, DB + class Update
1) OrderPage -- two new buttons that add to and remove from shopping
cart. A new DefaultListModel to hold our list of products that are read
from DB.

2) DBConnector -- Added helper methods to query DB (Read from Products
table, and write to order table). Still have to finish writing to
order_products table, (Implemented but not working as of yet, commented
out). New method that checks for Internet connectivity when a button is
pressed, and informs user if no Internet is available. (If no Internet,
then DB connection times out and displays SQL error, which is false)


3) Class updates -- removed some instance variables that are not needed.
Removed some setters that were mistakenly generated (set productNo etc)

--- Last big push ---