Class and DBConnector update
1) Class -- Account class constructor has been modified to assign number
of orders (Instead of assigning to zero).

2) DBConnector -- readAccount method has been modified to retrieve
number of orders from DB when signing in. A method has been added to
query the update on the orders field in DB. Another method has been
added to write the change to the DB.
Write order method has also been modified take in an extra parameter
(account), this is to execute the change after the order has been
made/placed (Keeping track and updating of orders in Account table).