Admin1 page update - FocusListener
Added FocusListeners to the two text fields.

Once focus is gained on a field, the field gets set to null
so the user can enter their details.

Once focus is lost, and text has been entered, the text remains.

If focus is lost and NO text was entered,
then the text field gets set to its original.