Admin2 page update
Edited the add button event method
to check the extra field (quantity) to
see if it is null or empty string.

When creating the product object, the
quantity was also used for construction.

Added the quantity as an element in the
String array when appending to the table model.

Edited getAddPanelDetails method name.