DBConnector update
Two new methods have been added.

1) Removing a Product from the database --

The prior existing remove product method would
only correctly remove the product if it was not
part of an order (Foreign key constraint).

The fix -- check if the product has been removed
successfully without error,

If not - execute a query to the table where the foreign
key constraint applies (order_product), and remove all
occurrences of the product,

Then run the query again to remove the product from the product
table (No more foreign key constraints).


Return true if both queries did not return
an SQL error.


2) Removing an Account from the database --

The exact same applies here, a foreign key constraint
prevented an account from being removed.

The fix -- the exact same approach was used.

This time, the order table was the cause of the error and
would be queried first to remove all occurrences of the account,

And again, a query to the account table is made to remove
the account (No more foreign key constraints).


Return true if both queries did not return
an SQL error.